requirements.txt
datasets 
langchain 
openai 
# tiktoken
PyPDF2
python-dotenv
langchain-astradb
langchain-openai